Change ssn to be int, not long

+++ src/services/tests/Demoulas.ProfitSharing.IntegrationTests/Reports/YearEnd/Update/INTERMEDIATE_VALUES.cs
@@ -4,7 +4,7 @@ public class INTERMEDIATE_VALUES
 {
     public long WS_FD_BADGE { get; set; } // PIC 9(7).
     public string? WS_FD_NAME { get; set; } // PIC X(25).
-    public long WS_FD_SSN { get; set; } // PIC 9(9)
+    public int WS_FD_SSN { get; set; } // PIC 9(9)
     public long WS_FD_PSN { get; set; } // PIC 9(11)
     public decimal WS_FD_AMT { get; set; } // PIC S9(8)V99
     public decimal WS_FD_DIST1 { get; set; } // PIC S9(8)V99
