+++ src/services/tests/Demoulas.ProfitSharing.IntegrationTests/Reports/YearEnd/Update/PAYBEN1_REC.cs
@@ -3,7 +3,7 @@
 public class PAYBEN1_REC
 {
     public long PYBEN_PSN1 { get; set; } // PIC 9(11).
-    public long PYBEN_PAYSSN1 { get; set; } // PIC 9(9).
+    public int PYBEN_PAYSSN1 { get; set; } // PIC 9(9).
     public string? PYBEN_TYPE1 { get; set; } // PIC X.
     public long PYBEN_PERCENT1 { get; set; } // PIC 9(3).
     public string? PYBEN_NAME1 { get; set; } // PIC X(25).
